refactor(event-creation): rename form types for clarity and update related schemas
