+++ packages/ui/event-creation/create-event-form.tsx
@@ -7,6 +7,8 @@ import { PlusCircle } from 'lucide-react'
 import { useFieldArray, useForm } from 'react-hook-form'
 import { z } from 'zod'
 
+import type { ContentResource } from '@coursebuilder/core/schemas/content-resource-schema'
+
 import {
 	Accordion,
 	AccordionContent,
@@ -31,22 +33,241 @@ import {
 	useToast,
 } from '../index'
 import AdvancedTagSelector from '../resources-crud/tag-selector'
-import {
-	MultipleEventsFormSchema,
-	type CreateEventFormProps,
-	type EventForm,
-	type EventSeriesForm,
-} from './types'
 
-type MultipleEvents = z.infer<typeof MultipleEventsFormSchema>
-
-interface Tag {
-	id: string
-	fields: {
-		label: string
-		name: string
-	}
+export type EventCreationResult = {
+	type: 'single' | 'series'
+	event?: ContentResource
+	eventSeries?: ContentResource
+	childEvents?: ContentResource[]
 }
+
+export type CreateEventFormProps = {
+	onSuccess: (result: EventCreationResult) => Promise<void>
+	createEvent: (data: EventFormData) => Promise<ContentResource>
+	createEventSeries: (data: EventSeriesFormData) => Promise<{
+		eventSeries: ContentResource
+		childEvents: ContentResource[]
+	}>
+	tags?: {
+		id: string
+		fields: {
+			label: string
+			name: string
+		}
+	}[]
+	allowMultipleEvents?: boolean
+	allowCoupons?: boolean
+	defaultTimezone?: string
+	defaultPrice?: number
+	defaultQuantity?: number
+	defaultCouponEnabled?: boolean
+	defaultCouponPercentageDiscount?:
+		| '1'
+		| '0.95'
+		| '0.9'
+		| '0.75'
+		| '0.6'
+		| '0.5'
+		| '0.4'
+		| '0.25'
+		| '0.1'
+}
+
+/**
+ * Schema for creating multiple events with shared product configuration
+ * Each event has its own title, dates, and tags, but they share price and quantity
+ * When multiple events are created, an event series is created to contain them
+ */
+export const MultipleEventsFormSchema = z
+	.object({
+		type: z.enum(['event', 'event-series']).default('event'),
+		eventSeries: z.object({
+			title: z.string().max(90).default(''),
+			description: z.string().optional(),
+			tagIds: z
+				.array(
+					z.object({
+						id: z.string(),
+						fields: z.object({
+							label: z.string(),
+							name: z.string(),
+						}),
+					}),
+				)
+				.nullish(),
+		}),
+		sharedFields: z.object({
+			price: z.number().min(0).nullish(),
+			quantity: z.number().min(-1).nullish(),
+		}),
+		coupon: z.object({
+			enabled: z.boolean().default(false),
+			percentageDiscount: z
+				.enum(['1', '0.95', '0.9', '0.75', '0.6', '0.5', '0.4', '0.25', '0.1'])
+				.optional(),
+			expires: z.date().optional(),
+		}),
+		events: z
+			.array(
+				z.object({
+					title: z.string().min(2).max(90),
+					startsAt: z.date().nullish(),
+					endsAt: z.date().nullish(),
+					tagIds: z
+						.array(
+							z.object({
+								id: z.string(),
+								fields: z.object({
+									label: z.string(),
+									name: z.string(),
+								}),
+							}),
+						)
+						.nullish(), // Tags are per-event, not shared
+				}),
+			)
+			.min(1),
+	})
+	.refine(
+		(data) => {
+			// Require series title if more than one event
+			if (data.events.length > 1 && data.eventSeries.title.trim().length < 2) {
+				return false
+			}
+			return true
+		},
+		{
+			message: 'Series title must be at least 2 characters',
+			path: ['eventSeries', 'title'],
+		},
+	)
+	.refine(
+		(data) => {
+			// Require percentage discount when coupon is enabled
+			if (data.coupon.enabled && !data.coupon.percentageDiscount) {
+				return false
+			}
+			return true
+		},
+		{
+			message: 'Percentage discount is required',
+			path: ['coupon', 'percentageDiscount'],
+		},
+	)
+	.refine(
+		(data) => {
+			// Require expiration date when coupon is enabled
+			if (data.coupon.enabled && !data.coupon.expires) {
+				return false
+			}
+			return true
+		},
+		{
+			message: 'Expiration date is required',
+			path: ['coupon', 'expires'],
+		},
+	)
+
+export type MultipleEventsForm = z.infer<typeof MultipleEventsFormSchema>
+
+/**
+ * Schema for adapter-level single event creation
+ * Includes database fields required by the adapter
+ */
+export const EventFormDataSchema = z.object({
+	type: z.literal('event'),
+	fields: z.object({
+		title: z.string().min(2).max(90),
+		startsAt: z.date().nullish(),
+		endsAt: z.date().nullish(),
+		description: z.string().optional(),
+		price: z.number().min(0).nullish(),
+		quantity: z.number().min(-1).nullish(),
+		state: z.string().optional(),
+		visibility: z.string().optional(),
+		slug: z.string().optional(),
+		tagIds: z
+			.array(
+				z.object({
+					id: z.string(),
+					fields: z.object({
+						label: z.string(),
+						name: z.string(),
+					}),
+				}),
+			)
+			.nullish(),
+	}),
+	coupon: z
+		.object({
+			enabled: z.boolean(),
+			percentageDiscount: z.string().optional(),
+			expires: z.date().optional(),
+		})
+		.optional(),
+})
+
+export type EventFormData = z.infer<typeof EventFormDataSchema>
+
+/**
+ * Schema for adapter-level event series creation
+ * Includes database fields required by the adapter
+ */
+export const EventSeriesFormDataSchema = z.object({
+	type: z.literal('event-series'),
+	eventSeries: z.object({
+		title: z.string().min(2).max(90),
+		description: z.string().optional(),
+		tagIds: z
+			.array(
+				z.object({
+					id: z.string(),
+					fields: z.object({
+						label: z.string(),
+						name: z.string(),
+					}),
+				}),
+			)
+			.nullish(),
+	}),
+	sharedFields: z.object({
+		price: z.number().min(0).nullish(),
+		quantity: z.number().min(-1).nullish(),
+	}),
+	childEvents: z
+		.array(
+			z.object({
+				type: z.literal('event'),
+				fields: z.object({
+					title: z.string().min(2).max(90),
+					startsAt: z.date().nullish(),
+					endsAt: z.date().nullish(),
+					description: z.string().optional(),
+					tagIds: z
+						.array(
+							z.object({
+								id: z.string(),
+								fields: z.object({
+									label: z.string(),
+									name: z.string(),
+								}),
+							}),
+						)
+						.nullish(),
+				}),
+			}),
+		)
+		.min(1),
+	coupon: z
+		.object({
+			enabled: z.boolean(),
+			percentageDiscount: z.string().optional(),
+			expires: z.date().optional(),
+		})
+		.optional(),
+})
+
+export type EventSeriesFormData = z.infer<typeof EventSeriesFormDataSchema>
 
 /**
  * Shared event creation form component that can be used across multiple apps
@@ -62,8 +283,10 @@ export function CreateEventForm({
 	defaultTimezone = 'America/Los_Angeles',
 	defaultPrice = 250,
 	defaultQuantity = 40,
+	defaultCouponEnabled = false,
+	defaultCouponPercentageDiscount = '0.25',
 }: CreateEventFormProps) {
-	const form = useForm<MultipleEvents>({
+	const form = useForm<MultipleEventsForm>({
 		resolver: zodResolver(MultipleEventsFormSchema),
 		defaultValues: {
 			type: 'event',
@@ -77,8 +300,8 @@ export function CreateEventForm({
 				quantity: defaultQuantity,
 			},
 			coupon: {
-				enabled: false,
-				percentageDiscount: '0.25',
+				enabled: defaultCouponEnabled,
+				percentageDiscount: defaultCouponPercentageDiscount,
 				expires: undefined,
 			},
 			events: [
@@ -99,7 +322,7 @@ export function CreateEventForm({
 
 	const { toast } = useToast()
 
-	const onSubmit = async (data: MultipleEvents) => {
+	const onSubmit = async (data: MultipleEventsForm) => {
 		try {
 			if (data.events.length === 1 && !data.eventSeries.title.trim()) {
 				// Single event - transform to adapter format (app wrapper adds system fields)
@@ -109,7 +332,7 @@ export function CreateEventForm({
 				}
 
 				const singleEventData: Omit<
-					EventForm,
+					EventFormData,
 					'createdById' | 'organizationId'
 				> = {
 					type: 'event',
@@ -140,7 +363,7 @@ export function CreateEventForm({
 			} else {
 				// Multiple events or event series - transform to adapter format
 				const eventSeriesData: Omit<
-					EventSeriesForm,
+					EventSeriesFormData,
 					'createdById' | 'organizationId'
 				> = {
 					type: 'event-series',
